% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_comp.R
\name{scale_color_comp}
\alias{scale_color_comp}
\title{ggplot2 color scale with complementary palette}
\usage{
scale_color_comp(...)
}
\arguments{
\item{...}{Arguments passed to comp_palette and scale_color_manual}
}
\description{
ggplot2 color scale with complementary palette
}
