% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_comp.R
\name{scale_fill_comp}
\alias{scale_fill_comp}
\title{ggplot2 fill scale with complementary palette}
\usage{
scale_fill_comp(...)
}
\arguments{
\item{...}{Arguments passed to comp_palette and scale_fill_manual}
}
\description{
ggplot2 fill scale with complementary palette
}
